+++ src/layout/Payment/SummaryPaymentComponent.tsx
@@ -6,10 +6,10 @@ import { Caption } from 'src/components/form/Caption';
 import { useLaxInstanceData } from 'src/features/instance/InstanceContext';
 import { Lang } from 'src/features/language/Lang';
 import { useCurrentLanguage } from 'src/features/language/LanguageProvider';
+import { usePaymentInformationQuery } from 'src/features/payment/PaymentInformationProvider';
 import { useInstanceIdParams } from 'src/hooks/useInstanceIdParams';
 import { getInstanceReferenceNumber } from 'src/layout/InstanceInformation/InstanceInformationComponent';
 import classes from 'src/layout/Payment/PaymentComponent.module.css';
-import { usePaymentInformationQuery } from 'src/layout/Payment/queries/usePaymentInformationQuery';
 import { PaymentDetailsTable } from 'src/layout/PaymentDetails/PaymentDetailsTable';
 import { formatDateLocale } from 'src/utils/formatDateLocale';
 import type { ISummaryComponent } from 'src/layout/Summary/SummaryComponent';
