+++ src/layout/PaymentDetails/PaymentDetailsTable.tsx
@@ -7,7 +7,7 @@ import cn from 'classnames';
 import { Caption } from 'src/components/form/Caption';
 import { Lang } from 'src/features/language/Lang';
 import classes from 'src/layout/PaymentDetails/PaymentDetailsTable.module.css';
-import type { OrderDetails } from 'src/layout/Payment/queries/types';
+import type { OrderDetails } from 'src/features/payment/types';
 
 type PaymentDetailsTableProps = {
   orderDetails?: OrderDetails;
